# api/openapi.yaml
openapi: 3.0.3
info:
  title: Mahzen API
  description: |
    Mahzen is a knowledge management platform that provides CRUD operations for
    entries and tags, keyword and semantic search, and user authentication.

    The HTTP API is served via Gin on port 8080. When TLS is configured, the
    server supports both HTTP/2 and HTTP/3 (QUIC).

    **Authentication**: Requests are authenticated via JWT Bearer tokens in the
    `Authorization` header. The auth middleware validates the token and sets
    `X-User-Id` internally. Public endpoints (register, login, refresh) do not
    require authentication.

    **Visibility**: Entries are either `public` (visible to all) or `private`
    (visible only to the owner). Listing and search endpoints enforce this at the
    database and search-engine level.

    **Pagination**: `limit` defaults to 20 when omitted or <= 0.
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development (Gin)
...
paths:
...
  /v1/auth/register:
    post:
      operationId: Register
      summary: Register a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/auth/login:
    post:
      operationId: Login
      summary: Login with email and password
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
  /v1/auth/refresh:
    post:
      operationId: RefreshToken
      summary: Refresh an access token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
  /v1/auth/logout:
    post:
      operationId: Logout
      summary: Invalidate a refresh token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
      responses:
        "200":
          description: Logged out successfully
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------------------------------------------------------------------------
  # Entries
  # ---------------------------------------------------------------------------
  /v1/entries:
    post:
      operationId: CreateEntry
      summary: Create a new entry
      tags: [Entries]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEntryRequest"
      responses:
        "200":
          description: Entry created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntryResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      operationId: ListEntries
      summary: List entries visible to the caller
      description: |
        Returns entries visible to the authenticated user. Public entries are
        always included; private entries are included only when they belong to
        the caller.

        When `path` is provided, results are filtered to entries at that path
        and all descendant paths (recursive). For example, `path=/notes` returns
        entries at `/notes`, `/notes/work`, `/notes/work/meeting`, etc.
      tags: [Entries]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PathFilter"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paginated list of entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEntriesResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/entries/{id}:
    get:
      operationId: GetEntry
      summary: Get a single entry by ID
      tags: [Entries]
      parameters:
        - $ref: "#/components/parameters/EntryID"
      responses:
        "200":
          description: Entry found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntryResponse"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      operationId: UpdateEntry
      summary: Update an existing entry
      tags: [Entries]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/EntryID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEntryRequest"
      responses:
        "200":
          description: Entry updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntryResponse"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      operationId: DeleteEntry
      summary: Delete an entry
      tags: [Entries]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/EntryID"
      responses:
        "200":
          description: Entry deleted successfully
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------------------------------------------------------------------------
  # Tags
  # ---------------------------------------------------------------------------
...
  /v1/search/keyword:
    get:
      operationId: KeywordSearch
      summary: Full-text keyword search over entries
      description: |
        Performs a keyword-based search via Typesense. Results respect visibility
        rules: public entries are always returned, private entries only when the
        caller owns them.
      tags: [Search]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/search/semantic:
    get:
      operationId: SemanticSearch
      summary: Semantic (vector) search over entries
      description: |
        Performs a semantic search by embedding the query via OpenAI
        (`text-embedding-3-small`, 1536 dimensions) and running a vector search
        in Typesense. Results respect visibility rules.
      tags: [Search]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------------------------------------------------------------------------
  # Users
  # ---------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ---------------------------------------------------------------------------
  # Parameters
  # ---------------------------------------------------------------------------
  parameters:
    EntryID:
      name: id
      in: path
      required: true
      description: Entry UUID
      schema:
        type: string
        format: uuid
...
    Limit:
      name: limit
      in: query
      required: false
      description: "Maximum number of items to return (default: 20)"
      schema:
        type: integer
        format: int32
        default: 20
        minimum: 1
    Offset:
      name: offset
      in: query
      required: false
      description: Number of items to skip
      schema:
        type: integer
        format: int32
        default: 0
        minimum: 0
    SearchQuery:
      name: query
      in: query
      required: true
      description: Search query string
      schema:
        type: string
    PathFilter:
      name: path
      in: query
      required: false
      description: |
        Filter entries by path prefix (recursive). Returns entries at the given
        path and all descendants. For example, `/notes` matches `/notes`,
        `/notes/work`, `/notes/work/meeting`, etc. Omit to list all entries.
      schema:
        type: string
        example: /notes

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
    Visibility:
      type: string
      enum:
        - public
        - private
      description: |
        Entry visibility level.
        - `public`: Visible to all users.
        - `private`: Visible only to the owner. This is the default.

    # -- Auth ----------------------------------------------------------------
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        display_name:
          type: string
        password:
          type: string
          minLength: 8
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    LogoutRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string

    # -- Entry -----------------------------------------------------------------
    Entry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        summary:
          type: string
          description: AI-generated summary (may be empty if summarizer is not enabled)
        path:
          type: string
          description: |
            Materialized path for hierarchical organization. Must start with `/`,
            no trailing slash (except root `/`). Segments allow `[a-zA-Z0-9._~-]`.
            Max segment length 255 chars, max total path 4096 bytes.
          example: /notes/work
        visibility:
          $ref: "#/components/schemas/Visibility"
        tags:
          type: array
          items:
            type: string
          description: Tag names associated with this entry
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateEntryRequest:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        path:
          type: string
          description: |
            Hierarchical path for the entry. Defaults to `/` (root) when omitted.
            Must start with `/`, no trailing slash. Segments allow `[a-zA-Z0-9._~-]`.
          example: /notes/work
        visibility:
          $ref: "#/components/schemas/Visibility"
        tag_ids:
          type: array
          items:
            type: string
            format: uuid
          description: Tag IDs to attach to the entry
    EntryResponse:
      type: object
      properties:
        entry:
          $ref: "#/components/schemas/Entry"
    UpdateEntryRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        path:
          type: string
          description: |
            New hierarchical path for the entry. If omitted or empty, the
            existing path is preserved.
          example: /notes/work
        visibility:
          $ref: "#/components/schemas/Visibility"
        tag_ids:
          type: array
          items:
            type: string
            format: uuid
    ListEntriesResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/Entry"
        total:
          type: integer
          format: int32
          description: Total number of entries matching the query

    # -- Tag -------------------------------------------------------------------
...
    SearchResult:
      type: object
      properties:
        entry_id:
          type: string
          format: uuid
        title:
          type: string
        snippet:
          type: string
          description: Relevant content excerpt
        score:
          type: number
          format: double
          description: Relevance score (keyword) or vector distance (semantic)
        highlights:
          type: array
          items:
            type: string
          description: Highlighted matching fragments
    SearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchResult"
        total:
          type: integer
          format: int32

    # -- User ------------------------------------------------------------------
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message

  # ---------------------------------------------------------------------------
  # Responses
  # ---------------------------------------------------------------------------
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthenticated:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"